lexicographically first palindrome using the input string.
Simple Approach: 1. Sort the string characters in alphabetical(ascending) order. 2. One be one find lexicographically next permutation of the given string. 3. The first permutation which is palindrome is the answer.

Efficient Approach: Properties for palindromic string: 1. If length of string is even, then the frequency of each character in the string must be even. 2. If the length is odd then there should be one character whose frequency is odd and all other chars must have even frequency and at-least one occurrence of the odd character must be present in the middle of the string.